% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Little bag of bootstraps on linear regression model}
\usage{
blblm(formula, data, m = 10, B = 5000, parallel = FALSE)
}
\arguments{
\item{formula}{A linear regression formula to be fitted to the bootstrapped data.}

\item{data}{A data frame containing the variables in the model.}

\item{m}{An integer specifying the number of subsets to split the data into.}

\item{B}{An integer specifying the number of times for the bootstrap procedure to be repeated.}

\item{parallel}{A boolean value indicating whether to use parallelization to compute fitted blblm model.
Default is set to FALSE. If parallelization is desired user must use future::plan() before use of blblm.}
}
\value{
Bag of little bootstraps linear regression model
}
\description{
Preform the bag of little of bootstraps method to compute the bootstrapped linear regression model.
}
\examples{
fit <- blblm(mpg ~ wt * hp, data = mtcars, m = 3, B = 100, parallel = FALSE)
}
