% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{blblm2-package}
\alias{blblm2}
\alias{blblm2-package}
\title{blblm2: Performs bag of little bootstraps on a linear regression model}
\description{
Improved performace of the original blblm package. Performs bag of little bootstraps on a linear regression model
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Linear Regression with Little Bag of Bootstraps
}
