% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{blblog}
\alias{blblog}
\title{Little bag of bootstraps on logistic regression model}
\usage{
blblog(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{A logistic regression formula to be fitted to the bootstrapped data.}

\item{data}{A data frame containing the variables in the model.}

\item{m}{An integer specifying the number of subsets to split the data into.}

\item{B}{An integer specifying the number of times for the bootstrap procedure to be repeated.}
}
\value{
Bag of little bootstrapped logistic regression model
}
\description{
Preform the bag of little bootstraps method to compute the bootstrapped logistic regression model.
}
\examples{
fit <- blblog(vs ~ wt + disp, data = mtcars, m = 3, B = 100)
}
