% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Calculate blblm coefficients}
\usage{
\method{coef}{blblm}(object, ...)
}
\arguments{
\item{object}{A fitted blblm model}

\item{...}{Additional arguments}
}
\value{
Coefficients of fitted blblm model.
}
\description{
Calculate the coefficients of the fitted blblm model.
}
