% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{coef.blblog}
\alias{coef.blblog}
\title{Calculate blblog coefficients}
\usage{
\method{coef}{blblog}(object, ...)
}
\arguments{
\item{object}{A fitted blblog model}

\item{...}{Additional arguments}
}
\value{
Coefficients of fitted blblog model.
}
\description{
Calculate the coefficients of the fitted blblog model.
}
