% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{confint.blblog}
\alias{confint.blblog}
\title{Confidence interval for blblog coefficients}
\usage{
\method{confint}{blblog}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted blblog model}

\item{parm}{The number of parameters in the fitted blblog model. Default is set to NULL.}

\item{level}{A numeric value indicating the desired level of confidence interval.}

\item{...}{Additional parameters}
}
\value{
A confidence interval for the blblog coefficients.
}
\description{
Compute confidence interval for the coefficients of a blblog model.
}
