% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1}
\alias{lm1}
\title{compute the linear regression estimates for a blb dataset}
\usage{
lm1(X, y, n)
}
\arguments{
\item{X}{matrix}

\item{y}{vector}

\item{n}{integer}
}
\description{
compute the linear regression estimates for a blb dataset
}
