% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{Compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{A linear regression formula to be fitted to the bootstrapped data.}

\item{data}{A data frame to be used to compute the estimates.}

\item{n}{An integer}

\item{B}{An integer specifying the number of repeats.}
}
\description{
Compute the estimates
}
