% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{log1}
\alias{log1}
\title{Compute the logistic regression estimates for a blb dataset}
\usage{
log1(X, y, n)
}
\arguments{
\item{X}{matrix}

\item{y}{vector}

\item{n}{integer}
}
\description{
Compute the logistic regression estimates for a blb dataset
}
