% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Compute predictions from fitted blblm model}
\usage{
\method{predict}{blblm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted blblm model.}

\item{new_data}{A data frame containing the the data to be predicted.}

\item{confidence}{A boolean parameter indicating if a confidence level is desired. Default is set to FALSE.}

\item{level}{A numeric value indicating the level of the confidence interval if desired.}

\item{...}{Additional arguments}
}
\description{
Predicts intercept of the linear regression model based on fitted blblm model using new data.
}
