% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{predict.blblog}
\alias{predict.blblog}
\title{Compute predictions from fitted blblog model}
\usage{
\method{predict}{blblog}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted blblog model.}

\item{new_data}{A data frame containing the the data to be predicted.}

\item{confidence}{A boolean parameter indicating if a confidence level is desired. Default is set to FALSE.}

\item{level}{A numeric value indicating the level of the confidence interval if desired.}

\item{...}{Additional arguments}
}
\description{
Compute predictions of coefficients of new data using a fitted blblog model.
}
