% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{print.blblm}
\alias{print.blblm}
\title{Print the blblm model}
\usage{
\method{print}{blblm}(x, ...)
}
\arguments{
\item{x}{A fitted blblm model}

\item{...}{Additional arguments}
}
\value{
The blblm model
}
\description{
Prints blblm fitted model to the console.
}
