% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblog.R
\name{print.blblog}
\alias{print.blblog}
\title{Print the blblog model}
\usage{
\method{print}{blblog}(x, ...)
}
\arguments{
\item{x}{A fitted blblog model}

\item{...}{Additional arguments}
}
\value{
The blblog model
}
\description{
Prints blblog fitted model to the console.
}
