% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Compute sigma of the blblm model}
\usage{
\method{sigma}{blblm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted blblm model.}

\item{confidence}{A boolean value indicating whether to return a confidince interval for the value of sigma.}

\item{level}{A numeric value indicating the confidence level. Used when confidence is set to TRUE.}

\item{...}{Additional arguments}
}
\value{
A numeric value indicating the value of sigma calculated from the blblm model.
}
\description{
Computes the estimated standard deviation of the blblm model.
}
