% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{split_data}
\alias{split_data}
\title{Split data into m parts of approximated equal sizes.}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{A data frame to be split.}

\item{m}{An integer specifying the number of subsets to split the data into.}
}
\description{
Split data into m parts of approximated equal sizes.
}
